requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
